+++ packages/core/src/defs/index.tsx
@@ -1,3 +1,4 @@
 export * from './ArrowMarker'
+export * from './BluntMarker'
 export * from './LinearGradient'
 export * from './types'